backend/requirements.txt
fastapi==0.136.1
uvicorn==0.46.0

pydantic==2.13.3
pydantic-settings==2.14.0

passlib[bcrypt]==1.7.4
bcrypt==3.2.2
python-jose==3.5.0

motor==3.7.1
pymongo==4.17.0

python-multipart==0.0.27
python-dotenv==1.2.2
email-validator==2.3.0

requests==2.33.1
httpx==0.28.1

pdfplumber==0.11.9
pdfminer.six==20251230
pypdfium2==5.7.1
pillow==12.2.0

openai==2.32.0
tqdm==4.67.3

numpy
cloudinary==1.44.2